Support python 2.7 (#200)
* Add diagram.

* Add py27 wheels.

* Fix error handling.

* Remove cruft.

* Redirect WORKSPACE.

* Fix names.

* Add test for python 2.7.

* Fix python 2.7 test.

* Fix clang-tidy.

* Cleanup

+++ ci/do_ci.sh
@@ -65,7 +65,7 @@ elif [[ "$1" == "clang_tidy" ]]; then
   setup_clang_toolchain
   bazel build \
         $BAZEL_OPTIONS \
-        -- //src/... //test/... //include/... //benchmark/...
+        -- //src/... //bridge/... //test/... //include/... //benchmark/...
   ./ci/gen_compilation_database.sh
   ./ci/fix_compilation_database.py
   ./ci/run_clang_tidy.sh |& tee /clang-tidy-result.txt
@@ -153,11 +153,22 @@ elif [[ "$1" == "plugin.test" ]]; then
   [[ "$NUM_SPANS" -eq "1" ]] || exit 1
   exit 0
 elif [[ "$1" == "python.wheel.test" ]]; then
-  python3 -m pip install /plugin/*.whl
+  # Python 3
+  python3 -m pip install /plugin/*cp3*.whl
   MOCK_SATELLITE_PORT=5000
   /plugin/mock_satellite $MOCK_SATELLITE_PORT &
   MOCK_SATELLITE_PID=$!
   python3 test/bridge/python/span_probe.py 127.0.0.1 $MOCK_SATELLITE_PORT
+  NUM_SPANS=`/plugin/mock_satellite_query --port $MOCK_SATELLITE_PORT num_spans`
+  kill $MOCK_SATELLITE_PID
+  [[ "$NUM_SPANS" -eq "1" ]] || exit 1
+
+  # Python 2.7
+  python -m pip install /plugin/*cp27mu*.whl
+  MOCK_SATELLITE_PORT=5000
+  /plugin/mock_satellite $MOCK_SATELLITE_PORT &
+  MOCK_SATELLITE_PID=$!
+  python test/bridge/python/span_probe.py 127.0.0.1 $MOCK_SATELLITE_PORT
   NUM_SPANS=`/plugin/mock_satellite_query --port $MOCK_SATELLITE_PORT num_spans`
   kill $MOCK_SATELLITE_PID
   [[ "$NUM_SPANS" -eq "1" ]] || exit 1
